Feature: display board and contents
In HTML, added a div to contain the board. Inside this div, there are
9 divs for the cells. And inside each 'cell' added a button to later use
them to mark them.

In CSS, added style to display the board using a grid. The body uses
flex to center both the h1 and the board. Also modified the font size of
the buttons to make it visible.

For Javascript, added 3 modules: one for the game board, another for the
logic to display stuff and the last one to handle the game's logic.
There is also one factory to make player objects.

The gameBoard has an 9 'O' array. The displayController has a function
to display the marks. The player factory just has a mark variable.