Feature: let player place marks on the board
In HTML, the buttons have data-attributes to easily identify them.

In Javascript, moved some objects. There are functions to start a new
round, update the h2 element, get the current player, udpate the current
player, add listeners to the buttons (e.g. board cells), etc.

All these functions let a player place a mark on a cell only once, like
the original game. It also updates the board array to later check if
a game ends in win or draw.