Feature: add logic to determine victory or draw
There is a function in gameBoard that returns true when the board is
not full. Also there are functions that return all the lines (rows,
columns and diagonals).

In gameLogic, there are functions to check if a player has filled a
line. Another function loops through every line to check if the player
has filled one line. Otherwise, if the gameBoard is full, then it's
a draw. Both options return a string.